ubatuba-event-tracker-backend/requirements.txt
anyio==3.7.1
click==8.1.8
fastapi==0.104.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
pydantic==1.10.13
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
httpx==0.26.0
pytest-asyncio==0.23.5